# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Shuttle Solution API for 42 Irbid Hackathon
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  student: "student",
  admin: "admin",
  driver: "driver",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role?: RegisterBodyRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  admin: "admin",
  driver: "driver",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface PickupPoint {
  id: number;
  name: string;
  lat: number;
  lng: number;
  routeOrder: number;
}

export type TripStatus = (typeof TripStatus)[keyof typeof TripStatus];

export const TripStatus = {
  pending: "pending",
  confirmed: "confirmed",
  canceled: "canceled",
} as const;

export type TripDirection = (typeof TripDirection)[keyof typeof TripDirection];

export const TripDirection = {
  to_school: "to_school",
  from_school: "from_school",
} as const;

export interface Trip {
  id: number;
  date: string;
  departureTime: string;
  status: TripStatus;
  totalSeats: number;
  bookedSeats: number;
  availableSeats: number;
  minBookingsToConfirm: number;
  direction: TripDirection;
  createdAt: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  canceled: "canceled",
} as const;

export interface Booking {
  id: number;
  userId: number;
  tripId: number;
  pickupPointId: number;
  status: BookingStatus;
  createdAt: string;
  trip?: Trip;
  pickupPoint?: PickupPoint;
  user?: User;
}

export interface CreateBookingBody {
  tripId: number;
  pickupPointId: number;
}

export type DriverTripDirection =
  (typeof DriverTripDirection)[keyof typeof DriverTripDirection];

export const DriverTripDirection = {
  to_school: "to_school",
  from_school: "from_school",
} as const;

export interface PickupStop {
  pickupPointId: number;
  pickupPointName: string;
  passengerCount: number;
  lat: number;
  lng: number;
  routeOrder: number;
}

export interface DriverTrip {
  id: number;
  date: string;
  departureTime: string;
  direction: DriverTripDirection;
  totalPassengers: number;
  pickupStops: PickupStop[];
}

export interface DashboardStats {
  totalBookingsToday: number;
  confirmedTripsToday: number;
  pendingTripsToday: number;
  totalStudents: number;
  totalTripsThisWeek: number;
  averageOccupancyRate: number;
}

export interface TripDemand {
  tripId: number;
  departureTime: string;
  direction: string;
  bookingCount: number;
  confirmedCount: number;
  pendingCount: number;
  availableSeats: number;
  status: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  trip_confirmed: "trip_confirmed",
  trip_canceled: "trip_canceled",
  booking_confirmed: "booking_confirmed",
  booking_canceled: "booking_canceled",
  general: "general",
} as const;

export interface Notification {
  id: number;
  userId: number;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export type GetTripsParams = {
  date?: string;
};

export type GetBookingsParams = {
  status?: GetBookingsStatus;
  date?: string;
};

export type GetBookingsStatus =
  (typeof GetBookingsStatus)[keyof typeof GetBookingsStatus];

export const GetBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  canceled: "canceled",
} as const;

export type GetAdminBookingsParams = {
  date?: string;
  status?: string;
  tripId?: number;
};

export type GetTripDemandParams = {
  date?: string;
};