# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Shuttle Solution API for 42 Irbid Hackathon
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get bookings (student sees own, admin sees all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, canceled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid booking request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/bookings:
    get:
      operationId: getAdminBookings
      tags: [bookings]
      summary: Get all bookings (admin only)
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: tripId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [users]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, admin, driver]
          default: student
      required:
        - name
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin, driver]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
    PickupPoint:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        routeOrder:
          type: integer
      required:
        - id
        - name
        - lat
        - lng
        - routeOrder
    Trip:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        departureTime:
          type: string
        status:
          type: string
          enum: [pending, confirmed, canceled]
        totalSeats:
          type: integer
        bookedSeats:
          type: integer
        availableSeats:
          type: integer
        minBookingsToConfirm:
          type: integer
        direction:
          type: string
          enum: [to_school, from_school]
        createdAt:
          type: string
      required:
        - id
        - date
        - departureTime
        - status
        - totalSeats
        - bookedSeats
        - availableSeats
        - minBookingsToConfirm
        - direction
        - createdAt
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        tripId:
          type: integer
        pickupPointId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, canceled]
        createdAt:
          type: string
        trip:
          $ref: "#/components/schemas/Trip"
        pickupPoint:
          $ref: "#/components/schemas/PickupPoint"
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - tripId
        - pickupPointId
        - status
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        tripId:
          type: integer
        pickupPointId:
          type: integer
      required:
        - tripId
        - pickupPointId
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [trip_confirmed, trip_canceled, booking_confirmed, booking_canceled, general]
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - message
        - type
        - isRead
        - createdAt