# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Shuttle Solution API for 42 Irbid Hackathon
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { PickupPoint } from "./pickupPoint";
import type { Trip } from "./trip";
import type { User } from "./user";

export interface Booking {
  id: number;
  userId: number;
  tripId: number;
  pickupPointId: number;
  status: BookingStatus;
  createdAt: string;
  trip?: Trip;
  pickupPoint?: PickupPoint;
  user?: User;
}