# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Shuttle Solution API for 42 Irbid Hackathon
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  tripId: number;
  pickupPointId: number;
}